refactor: move new Meteor startup stuff to separate startNodeApp function
